cli: Split the executable and the library parts
After some reading and looking at main.rs and command.rs, and
considering our pursue to better split modules in this project,
splitting the executable and the library componentes seems like a good
idea, as well as separating the "lib-like" types from the specific
CommandT implementation.
